text-to-image-openai/requirements.txt
openai 
gradio 
argparse
pillow